création dossier user

+++ src/View/view-post.php
@@ -42,14 +42,18 @@
                             <br />
                             <b><?= date("d/m/Y - H:i", $post['post_timestamp']) ?></b>
                             <br />
-                            <span class="post-edit">
-                                <button>
-                                    <i class="fa-solid fa-pen-to-square"></i> éditer
-                                </button>
-                                <a href="./controller-delete_posts.php?post_id=<?= $post['post_id'] ?>">
-                                    <button><i class="fa-solid fa-square-xmark"></i> supprimer</button>
-                                </a>
-                            </span>
+                            <?php if ($post['user_id'] == $_SESSION['user_id']) { ?>
+                                <span class="post-edit">
+                                    <a href="#">
+                                        <button>
+                                            <i class="fa-solid fa-pen-to-square"></i> éditer
+                                        </button>
+                                    </a>
+                                    <a href="./controller-delete_posts.php?post_id=<?= $post['post_id'] ?>">
+                                        <button><i class="fa-solid fa-square-xmark"></i> supprimer</button>
+                                    </a>
+                                </span>
+                            <?php } ?>
                         </p>
                     </div>
                     <div class="post-comments">
@@ -61,12 +65,16 @@
                                 </p>
                                 <span>
                                     <?= date("d/m/Y - H:i", $comment['com_timestamp']) ?>
-                                    <button>
-                                        <i class="fa-solid fa-pen"></i> éditer
-                                    </button>
-                                    <a href="./controller-delete_comments.php?post_id=<?= $comment['post_id'] ?>&com_id=<?= $comment['com_id'] ?>">
-                                        <button><i class="fa-solid fa-delete-left"></i> supprimer</button>
-                                    </a>
+                                    <?php if ($comment['user_id'] == $_SESSION['user_id']) { ?>
+                                        <a href="#">
+                                        <button>
+                                            <i class="fa-solid fa-pen"></i> éditer
+                                        </button>
+                                        </a>
+                                        <a href="./controller-delete_comments.php?post_id=<?= $comment['post_id'] ?>&com_id=<?= $comment['com_id'] ?>">
+                                            <button><i class="fa-solid fa-delete-left"></i> supprimer</button>
+                                        </a>
+                                    <?php } ?>
                                 </span>
                             </div>
                         <?php } ?>
